factor out HitCounter
